[codex] split data fitting facade helpers (#2615)
* refactor: split data fitting facade helpers

* style: format data fitting helper modules

* chore: trigger CI

---------
